+++ docs/changelog.rst
@@ -23,6 +23,7 @@ Version 1.7.0
 - Allow ``ResourceProtector`` decorator to be used without parentheses. :issue:`604`
 - Implement RFC9700 PKCE downgrade countermeasure.
 - Set ``User-Agent`` header when fetching server metadata and JWKs. :issue:`704`
+- RFC7523 accepts the issuer URL as a valid audience. :issue:`730`
 
 Upgrade Guide: :ref:`joserfc_upgrade`.